LayoutTests/imported/w3c/web-platform-tests/content-security-policy/gen/top.meta/worker-src-none/worklet-audio-import-data.https-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'worker-src'.

Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/worker.py?redirection=keep-origin&action=purge&key=61260440-880a-4bea-90f5-fa1d31e7bcc1&path=/mixed-content
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/worker.py?redirection=no-redirect&action=purge&key=b88c8780-10c1-4495-bd3c-8a7bc645b502&path=/mixed-content
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/worker.py?redirection=swap-origin&action=purge&key=72b1b1bb-2b00-475b-9384-d397f3018f7f&path=/mixed-content
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/worker.py?action=purge&key=8614abc9-bf8d-4753-96cd-9d8b7a75bd0c&path=%2Fmixed-content

FAIL Content Security Policy: Expects allowed for worklet-audio-import-data to cross-https origin and keep-origin redirection from https context. promise_test: Unhandled rejection with value: object "AbortError: Cross-origin script load denied by Cross-Origin Resource Sharing policy."
PASS Content Security Policy: Expects allowed for worklet-audio-import-data to cross-https origin and keep-origin redirection from https context.: securitypolicyviolation
FAIL Content Security Policy: Expects allowed for worklet-audio-import-data to cross-https origin and no-redirect redirection from https context. promise_test: Unhandled rejection with value: object "AbortError: Cross-origin script load denied by Cross-Origin Resource Sharing policy."
PASS Content Security Policy: Expects allowed for worklet-audio-import-data to cross-https origin and no-redirect redirection from https context.: securitypolicyviolation
FAIL Content Security Policy: Expects allowed for worklet-audio-import-data to cross-https origin and swap-origin redirection from https context. promise_test: Unhandled rejection with value: object "AbortError: Cross-origin script load denied by Cross-Origin Resource Sharing policy."
PASS Content Security Policy: Expects allowed for worklet-audio-import-data to cross-https origin and swap-origin redirection from https context.: securitypolicyviolation
PASS Content Security Policy: Expects allowed for worklet-audio-import-data to same-https origin and keep-origin redirection from https context.
PASS Content Security Policy: Expects allowed for worklet-audio-import-data to same-https origin and keep-origin redirection from https context.: securitypolicyviolation
PASS Content Security Policy: Expects allowed for worklet-audio-import-data to same-https origin and no-redirect redirection from https context.
PASS Content Security Policy: Expects allowed for worklet-audio-import-data to same-https origin and no-redirect redirection from https context.: securitypolicyviolation
FAIL Content Security Policy: Expects allowed for worklet-audio-import-data to same-https origin and swap-origin redirection from https context. promise_test: Unhandled rejection with value: object "AbortError: Importing a module script is canceled."
PASS Content Security Policy: Expects allowed for worklet-audio-import-data to same-https origin and swap-origin redirection from https context.: securitypolicyviolation
